+++ src/main/java/ladylib/client/command/ShaderReloadCommand.java
@@ -1,10 +1,14 @@
 package ladylib.client.command;
 
-import ladylib.client.shader.ShaderUtil;
-import net.minecraft.client.Minecraft;
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.util.text.Style;
+import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraftforge.client.resource.VanillaResourceType;
+import net.minecraftforge.common.ForgeModContainer;
+import net.minecraftforge.fml.client.FMLClientHandler;
 
 public class ShaderReloadCommand extends CommandBase {
     @Override
@@ -19,6 +23,9 @@ public class ShaderReloadCommand extends CommandBase {
 
     @Override
     public void execute(MinecraftServer server, ICommandSender sender, String[] args) {
-        ShaderUtil.loadShaders(Minecraft.getMinecraft().getResourceManager());
+        if (!ForgeModContainer.selectiveResourceReloadEnabled) {
+            sender.sendMessage(new TextComponentTranslation("ladylib.warning.no_selective_reload").setStyle(new Style().setColor(TextFormatting.YELLOW)));
+        }
+        FMLClientHandler.instance().refreshResources(VanillaResourceType.SHADERS);
     }
 }