+++ src/main/java/ladylib/reflection/Getter.java
@@ -1,0 +1,9 @@
+package ladylib.reflection;
+
+import java.lang.invoke.MethodHandle;
+
+public class Getter<T, R> extends LLMethodHandle.LLMethodHandle0<T, R> {
+    Getter(MethodHandle methodHandle, String name) {
+        super(methodHandle, name);
+    }
+}