+++ src/main/java/ladylib/reflection/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Reflection helpers
+ * TODO move {@link ladylib.misc.ReflectionFailedException} and {@link ladylib.misc.ReflectionUtil} here
+ */
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+package ladylib.reflection;
+
+import mcp.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;