+++ changelog.md
@@ -1,3 +1,20 @@
+##### Version 1.12.2-2.4.0 - BUILT
+Shaders:
+- Overhauled the screen/post-process shader system
+- Deprecated old screen shaders methods
+- Made the /ladylib_shader_reload command reload shader resources using Forge's new selective resource reloading
+- Dynamic lights now use a post process shader
+
+Reflection:
+- Added generic typed method handles
+- Added a new reflection helper that uses said typed method handles
+
+Kotlin:
+- Added a Kotlin module containing a few methods smoothing the experience for Forgelin fans
+
+Bugfixes:
+- Fixed the """Enhanced""" Event Subscriber not registering any class after an invalid one
+
 ##### Version 1.12.2-2.3.2 - BUILT
 Fixed automatic capability registration behaving opposite to what the javadoc says
 
@@ -11,7 +28,7 @@ Miscellaneous:
 Bugfixes:
 - Fixed automatic capability registrations when defaults are used
 
-##### Version 1.12.2-2.3 - BUILT
+##### Version 1.12.2-2.3.0 - BUILT
 Main changes:
 - Implemented cheap dynamic lights as an alternative to Albedo